Moved drop fumble to Item
Moved fumble out of drop.
Moved fumble into item, called after drop.
