+++ Content.Server/GameObjects/Components/GUI/ServerHandsComponent.cs
@@ -248,11 +248,7 @@ namespace Content.Server.GameObjects
 
             // TODO: The item should be dropped to the container our owner is in, if any.
             item.Owner.Transform.GridPosition = Owner.Transform.GridPosition;
-            if (item.Owner.TryGetComponent<PhysicsComponent>(out var physicsComponent))
-            {
-                physicsComponent.LinearVelocity += RandomOffset();
-            }
-
+            item.Fumble();
             if (item.Owner.TryGetComponent<SpriteComponent>(out var spriteComponent))
             {
                 spriteComponent.RenderOrder = item.Owner.EntityManager.CurrentTick.Value;
@@ -260,16 +256,6 @@ namespace Content.Server.GameObjects
 
             Dirty();
             return true;
-        }
-
-        private Vector2 RandomOffset()
-        {
-            return new Vector2(RandomOffset(), RandomOffset());
-            float RandomOffset()
-            {
-                var size = 15.0F;
-                return (new Random().NextFloat() * size) - size / 2;
-            }
         }
 
         public bool Drop(IEntity entity)
